+++ wponion-init.php
@@ -3,8 +3,8 @@
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
-defined( 'WPONION_VERSION' ) or define( 'WPONION_VERSION', '1.3.8' );
-defined( 'WPONION_CDN_VERSION' ) or define( 'WPONION_CDN_VERSION', '1.3.7' );
+defined( 'WPONION_VERSION' ) or define( 'WPONION_VERSION', '1.3.9' );
+defined( 'WPONION_CDN_VERSION' ) or define( 'WPONION_CDN_VERSION', '1.3.8' );
 defined( 'WPONION_CDN_URL' ) or define( 'WPONION_CDN_URL', 'https://cdn.jsdelivr.net/gh/wponion/vendor-support' );
 defined( 'WPONION_NAME' ) or define( 'WPONION_NAME', 'WPOnion' );
 defined( 'WPONION_FILE' ) or define( 'WPONION_FILE', plugin_dir_path( __FILE__ ) . 'index.php' );